requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
urllib3==2.6.3
